Complete Azure Function deployment with the use of Fileshare as the test project cannot run as single exe in .net 6. Add Test filter logic.

+++ GraderFunctionApp/GraderFunction.cs
@@ -40,6 +40,8 @@ namespace GraderFunctionApp
         Azure Credentials<br/>
         <textarea name='credentials' required  rows='15' cols='100'></textarea>
         <br/>
+        NUnit Test filter<br/>
+        <input type='text' id='filter' name='filter' /><br/>
         <button type='submit'>Run Test</button>
     </form>
     <footer>
@@ -59,6 +61,7 @@ namespace GraderFunctionApp
                 else
                 {
                     string credentials = req.Query["credentials"];
+                    string filter = req.Query["filter"];
 
                     string xml;
                     if (req.Query.ContainsKey("trace"))
@@ -66,12 +69,12 @@ namespace GraderFunctionApp
                         string trace = req.Query["trace"];
                         var email = ExtractEmail(trace);
                         log.LogInformation("start:" + trace);
-                        xml = await RunUnitTestProcess(context, log, credentials, email);
+                        xml = await RunUnitTestProcess(context, log, credentials, email, filter);
                         log.LogInformation("end:" + trace);
                     }
                     else
                     {
-                        xml = await RunUnitTestProcess(context, log, credentials);
+                        xml = await RunUnitTestProcess(context, log, credentials, "Anonymous", filter);
                     }
                     return new ContentResult { Content = xml, ContentType = "application/xml", StatusCode = 200 };
                 }
@@ -81,6 +84,7 @@ namespace GraderFunctionApp
             {
                 log.LogInformation("POST Request");
                 string credentials = req.Form["credentials"];
+                string filter = req.Form["filter"];
                 if (credentials == null)
                 {
                     return new ContentResult
@@ -90,24 +94,29 @@ namespace GraderFunctionApp
                         StatusCode = 422
                     };
                 }
-                var xml = await RunUnitTestProcess(context, log, credentials);
+                var xml = await RunUnitTestProcess(context, log, credentials, "Anonymous", filter);
 
                 return new ContentResult { Content = xml, ContentType = "application/xml", StatusCode = 200 };
             }
 
             return new OkObjectResult("ok");
         }
-        //as
-        private static async Task<string> RunUnitTestProcess(ExecutionContext context, ILogger log, string credentials, string trace = "NoTrace")
+
+        private static async Task<string> RunUnitTestProcess(ExecutionContext context, ILogger log, string credentials, string trace, string filter)
         {
             var tempDir = GetTemporaryDirectory(trace);
             var tempCredentialsFilePath = Path.Combine(tempDir, "azureauth.json");
 
             await File.WriteAllLinesAsync(tempCredentialsFilePath, new string[] { credentials });
 
-            string workingDirectoryInfo = context.FunctionAppDirectory;
-            string exeLocation = Path.Combine(context.FunctionAppDirectory, "AzureProjectGrader.exe");
+            string workingDirectoryInfo = Environment.ExpandEnvironmentVariables(@"%HOME%\data\Functions\Tests");
+            string exeLocation = Path.Combine(workingDirectoryInfo, "AzureProjectTest.exe");
+            log.LogInformation("exeLocation: " + exeLocation);
 
+            if(string.IsNullOrEmpty(filter))
+                filter = "test==AzureProjectTest";
+
+            log.LogInformation($@"{tempCredentialsFilePath} {tempDir} {trace} {filter}");
             try
             {
                 using var process = new Process();
@@ -115,11 +124,11 @@ namespace GraderFunctionApp
                 {
                     WorkingDirectory = workingDirectoryInfo,
                     FileName = exeLocation,
-                    Arguments = $@"{tempCredentialsFilePath} {tempDir} {trace}",
+                    Arguments = $@"{tempCredentialsFilePath} {tempDir} {trace} {filter}",
                     UseShellExecute = false,
                     CreateNoWindow = true,
                     RedirectStandardOutput = true,
-                    RedirectStandardError = true
+                    RedirectStandardError = true,
                 };
 
                 process.StartInfo = info;
@@ -129,53 +138,55 @@ namespace GraderFunctionApp
                 log.LogInformation("Process start.");
                 var output = new StringBuilder();
                 var error = new StringBuilder();
-                using (AutoResetEvent outputWaitHandle = new AutoResetEvent(false))
-                using (AutoResetEvent errorWaitHandle = new AutoResetEvent(false))
+                using AutoResetEvent outputWaitHandle = new(false);
+                using AutoResetEvent errorWaitHandle = new(false);
+                process.OutputDataReceived += (sender, e) =>
                 {
-                    process.OutputDataReceived += (sender, e) =>
+                    if (e.Data == null)
                     {
-                        if (e.Data == null)
-                        {
-                            outputWaitHandle.Set();
-                        }
-                        else
-                        {
-                            output.AppendLine(e.Data);
-                        }
-                    };
-                    process.ErrorDataReceived += (sender, e) =>
-                    {
-                        if (e.Data == null)
-                        {
-                            errorWaitHandle.Set();
-                        }
-                        else
-                        {
-                            error.AppendLine(e.Data);
-                        }
-                    };
-                    process.Start();
-                    process.BeginOutputReadLine();
-                    process.BeginErrorReadLine();
-
-                    const int timeout = 5 * 60 * 1000;
-                    if (process.WaitForExit(timeout) &&
-                        outputWaitHandle.WaitOne(timeout) &&
-                        errorWaitHandle.WaitOne(timeout))
-                    {
-                        // Process completed. Check process.ExitCode here.
-                        log.LogInformation("Process Ended.");
-
-                        log.LogInformation(error.ToString());
-                        var xml = await File.ReadAllTextAsync(Path.Combine(tempDir, "TestResult.xml"));
-                        Directory.Delete(tempDir, true);
-                        return xml;
+                        outputWaitHandle.Set();
                     }
                     else
                     {
-                        // Timed out.
-                        log.LogInformation("Process Timed out.");
+                        output.AppendLine(e.Data);
                     }
+                };
+                process.ErrorDataReceived += (sender, e) =>
+                {
+                    if (e.Data == null)
+                    {
+                        errorWaitHandle.Set();
+                    }
+                    else
+                    {
+                        error.AppendLine(e.Data);
+                    }
+                };
+                process.Start();
+                process.BeginOutputReadLine();
+                process.BeginErrorReadLine();
+
+                const int timeout = 5 * 60 * 1000;
+                if (process.WaitForExit(timeout) &&
+                    outputWaitHandle.WaitOne(timeout) &&
+                    errorWaitHandle.WaitOne(timeout))
+                {
+                    // Process completed. Check process.ExitCode here.
+                    log.LogInformation("Process Ended.");
+                    //log.LogInformation(output.ToString());
+
+                    var errorLog = error.ToString();
+                    log.LogError(errorLog);
+                    if (!string.IsNullOrEmpty(errorLog)) return null;
+
+                    var xml = await File.ReadAllTextAsync(Path.Combine(tempDir, "TestResult.xml"));
+                    Directory.Delete(tempDir, true);
+                    return xml;
+                }
+                else
+                {
+                    // Timed out.
+                    log.LogInformation("Process Timed out.");
                 }
             }
             catch (Exception ex)
